requirements.txt
aiocache==0.12.3
aiomcache==0.8.2
aioredis==2.0.1
annotated-types==0.7.0
anyio==4.8.0
async-timeout==5.0.1
certifi==2025.1.31
click==8.1.8
dnspython==2.7.0
email_validator==2.2.0
exceptiongroup==1.2.2
fastapi==0.115.5
fastapi-cache2==0.2.2
fastapi-cli==0.0.7
h11==0.14.0
httpcore==1.0.7
httptools==0.6.4
httpx==0.28.1
idna==3.10
Jinja2==3.1.5
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
pendulum==3.0.0
pydantic==2.10.6
pydantic-settings==2.6.1
pydantic_core==2.27.2
Pygments==2.19.1
pymemcache==4.0.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.20
PyYAML==6.0.2
redis==5.2.1
rich==13.9.4
rich-toolkit==0.13.2
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
starlette==0.41.3
time-machine==2.16.0
typer==0.15.2
typing_extensions==4.12.2
tzdata==2025.1
uvicorn==0.34.0
uvloop==0.21.0
watchfiles==1.0.4
websockets==15.0
